Rationalize some extension methods
