+++ src/commonMain/kotlin/io/klogging/impl/KloggerImpl.kt
@@ -21,6 +21,7 @@ package io.klogging.impl
 import io.klogging.Klogger
 import io.klogging.Level
 import io.klogging.context.LogContext
+import io.klogging.eventFrom
 import io.klogging.events.LogEvent
 import io.klogging.events.currentContext
 import io.klogging.events.timestampNow
@@ -30,7 +31,6 @@ import kotlin.coroutines.coroutineContext
 public class KloggerImpl(
     override val name: String,
 ) : Klogger {
-
     override suspend fun emitEvent(level: Level, exception: Exception?, event: Any?) {
         val eventToLog = eventFrom(level, exception, event, contextItems())
         Logging.sendEvent(eventToLog)
