+++ src/jvmTest/kotlin/io/klogging/events/LogEventTest.kt
@@ -18,12 +18,16 @@
 
 package io.klogging.events
 
+import io.klogging.Level.DEBUG
+import io.klogging.Level.INFO
 import io.klogging.context.logContext
 import io.klogging.logger
+import io.klogging.randomString
 import io.klogging.savedEvents
 import io.klogging.waitForDispatch
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.maps.shouldContain
+import io.kotest.matchers.maps.shouldContainAll
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.launch
 
@@ -45,6 +49,38 @@ class LogEventTest : DescribeSpec({
                     waitForDispatch()
                     saved.first().items shouldContain ("colour" to "white")
                 }
+            }
+        }
+    }
+
+    describe("LogEvent.copyWith() extension function") {
+        it("appends supplied items to those already in the event") {
+            val id = randomString()
+            val run = randomString()
+            val event = LogEvent(
+                logger = "CopyWithTest",
+                level = INFO,
+                message = randomString(),
+                items = mapOf("id" to id),
+            )
+            with(event.copyWith(DEBUG, null, mapOf("run" to run))) {
+                items shouldContainAll mapOf(
+                    "id" to id,
+                    "run" to run,
+                )
+            }
+        }
+        it("ignores any supplied items with the same keys as those already present") {
+            val run1 = randomString()
+            val run2 = randomString()
+            val event = LogEvent(
+                logger = "CopyWithTest",
+                level = INFO,
+                message = randomString(),
+                items = mapOf("run" to run1),
+            )
+            with(event.copyWith(DEBUG, null, mapOf("run" to run2))) {
+                items shouldBe mapOf("run" to run1)
             }
         }
     }
